Split account page into visual and logical components
